moving to a single screen : eventscreen
