+++ app/src/main/java/com/monkeyteam/chimpagne/ui/DetailScreenSheet.kt
@@ -21,7 +21,6 @@ import androidx.compose.foundation.rememberScrollState
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material.icons.automirrored.rounded.Login
-import androidx.compose.material.icons.rounded.Login
 import androidx.compose.material.icons.rounded.QrCodeScanner
 import androidx.compose.material3.Button
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -126,7 +125,7 @@ fun DetailScreenSheet(
               },
               { showToast(stringResFailure) })
 
-          navObject.clearAndNavigateTo(Route.VIEW_DETAIL_EVENT_SCREEN + "/${it.id})", false)
+          navObject.clearAndNavigateTo(Route.EVENT_SCREEN + "/${it.id})", false)
         }
       }
