+++ app/src/main/java/com/monkeyteam/chimpagne/ui/navigation/NavigationActions.kt
@@ -14,7 +14,7 @@ object Route {
   const val LOADING = "loading"
   const val EVENT_CREATION_SCREEN = "EVENT_CREATION_SCREEN"
   const val MY_EVENTS_SCREEN = "myEvents"
-  const val VIEW_DETAIL_EVENT_SCREEN = "viewDetailEventScreen"
+  const val EVENT_SCREEN = "viewDetailEventScreen"
   const val ONLINE_EVENT_VIEW = "onlineEventView"
   const val LOADING_LOGIN = "loadingLogin"
   const val MANAGE_STAFF_SCREEN = "manageStaffScreen"
